main/requirements.txt
asgiref==3.6.0
async-generator==1.10
async-timeout==3.0.1
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-payments==1.0.0
django-payments-razorpay==1.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
django-twilio==0.14.3
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
filelock==3.8.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
outcome==1.2.0
phonenumbers==8.12.55
phonenumberslite==8.13.4
Pillow==9.4.0
platformdirs==2.6.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
wsproto==1.2.0
zope.interface==5.5.2
